+++ README.md
@@ -26,7 +26,7 @@
   - [x] 0~5세, 65세~ -> 0원 부과
 
 ## 2단계 리팩토링 목록
-- [ ] 도메인 패키지 분리
+- [x] 도메인 패키지 분리
 - [x] PathService 메서드 분리
 - [x] 경로 조회 프로세스 책임 분리
 - [ ] 접근 제한자 확인 및 final 선언
