+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,7 +1,7 @@
 package wooteco.subway.dao;
 
 import java.util.List;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public interface LineDao {
