+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -1,7 +1,7 @@
 package wooteco.subway.dao;
 
 import java.util.List;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.secion.Section;
 
 public interface SectionDao {
