+++ src/main/java/wooteco/subway/dao/SectionJdbcDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.secion.Section;
 import wooteco.subway.domain.Station;
 
 @Repository
