+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.Sections;
+import wooteco.subway.domain.secion.Section;
+import wooteco.subway.domain.secion.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.SectionRequest;
 import wooteco.subway.exception.AddSectionException;
