+++ src/test/java/wooteco/subway/dao/LineJdbcDaoTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public class LineJdbcDaoTest extends DaoImplTest {
