+++ src/test/java/wooteco/subway/domain/section/SectionTest.java
@@ -1,9 +1,11 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.Station;
+import wooteco.subway.domain.secion.Section;
 import wooteco.subway.exception.PositiveDigitException;
 
 class SectionTest {
